Cleaned up highcharts plugin
