+++ packages/perspective-viewer-highcharts/test/results/results.json
@@ -42,7 +42,7 @@
     "heatmap.html/pivots by two rows.": "7808b22cad7cd787d2d9fdd70d21c912",
     "heatmap.html/pivots by a row and a column.": "98b0e9b447ecea7c427e4b08f14faec2",
     "heatmap.html/pivots by two rows and two columns.": "e9280f2060314559f512e6ac193be3ac",
-    "heatmap.html/sorts by a hidden column.": "e08ceb73f3dc477387252bff1d2f9e31",
+    "heatmap.html/sorts by a hidden column.": "cf92701b055e836148a5f4d332cec342",
     "heatmap.html/sorts by a numeric column.": "c87bee0e512b1a0dfab03255654db2cb",
     "heatmap.html/filters by a numeric column.": "84d54b0b81255811d339831774d6c904",
     "heatmap.html/sorts by an alpha column.": "431bf132963054f6c96a82eadfe14652",
